add ability to install themes and plugins as submodules.

+++ .scripts/build.sh
@@ -29,29 +29,71 @@ fi
 # =======================
 fetch_git(){
   REPOSITORY=$1
-  THEMENAME=$(basename "$REPOSITORY")
-  THEMENAME="${THEMENAME%.*}" #removing extension
-  TARGETDIR="${ROOT}/wp-content/themes/${THEMENAME}"
-
-  mkdir $TARGETDIR
-
+  NAME=$2
+  TARGETSUB=$3
+  TARGETDIR="${ROOT}${TARGETSUB}${NAME}"
   if [ ! -z "${REPOSITORY}" ]; then
-    echo "Local repository found. Downloading..."
-    git clone $REPOSITORY $TARGETDIR
-    build_theme $THEMENAME
+    read -p "Would you like to clone or install as submodule (c/s)?" yn
+    case $yn in
+        [Cc]* )
+          mkdir $TARGETDIR
+          git clone $REPOSITORY $TARGETDIR
+          break;;
+        [Ss]* )
+          echo "#Excluding submodule '${NAME}'\n!$TARGETSUB${NAME}" >> .gitignore
+          git submodule add $REPOSITORY ".${TARGETSUB}${NAME}"
+          break;;
+        * ) echo "Please answer c or s.";;
+    esac
   else
     echo "${RED}Trying to fetch an invalid git url. exiting..."
     exit
   fi
 }
 
+fetch_theme(){
+  while true; do
+    read -p "\nWould you like to install a(nother) theme?" yn
+    case $yn in
+      [Yy]* )
+        shopt -s nullglob
+        WPTHEMES=('git repository URL');
+        WPTHEMES+=($(ls -d $THEMES | xargs -n1 basename))
+        echo "\nPlease select the theme to install:\n"
+        select opt in "${WPTHEMES[@]}"
+        do
+          if [[ $opt = 'git repository URL' ]]; then
+            read -p "Write your theme repository URL: " REPOSITORY
+            THEMENAME=$(basename "$REPOSITORY")
+            THEMENAME="${THEMENAME%.*}" #removing extension
+            fetch_git $REPOSITORY $THEMENAME '/wp-content/themes/'
+            build_theme $THEMENAME
+          else
+            build_theme $opt
+          fi
+        done
+        break;;
+      [Nn]* ) exit;;
+      * ) echo "Please answer yes or no.";;
+    esac
+  done
+}
+
 build_theme(){
-  THEMEDIR="$ROOT/wp-content/themes/$1"
-  echo  "${GREEN}Building $1 theme…${NC}"
-  cd "${THEMEDIR}"
-  composer install --no-interaction $OPT_COMPOSER
-  npm install
-  npm run build --if-present
+  THEMENAME=$1
+  read -p "Would you like to build '${PLUGINNAME}'?" yn
+  case $yn in
+    [Yy]* )
+      THEMEDIR="$ROOT/wp-content/themes/$THEMENAME"
+      echo  "${GREEN}Building $1 theme…${NC}"
+      cd "${THEMEDIR}"
+      composer install --no-interaction $OPT_COMPOSER
+      npm install
+      npm run build --if-present
+      break;;
+    [Nn]* ) exit;;
+    * ) echo "Please answer yes or no.";;
+  esac
   remove_other_themes $1;
 }
 
@@ -59,14 +101,43 @@ remove_other_themes(){
   while true; do
     read -p "Would you like to remove the unused themes?" yn
     case $yn in
+      [Yy]* )
+        CURRENT=$1
+        WPTHEMES=($THEMES)
+        for THEME in "${WPTHEMES[@]}"; do
+          if [ $(basename $THEME) != $CURRENT ]; then
+            rm -rf $THEME
+          fi
+        done
+        break;;
+      [Nn]* ) exit;;
+      * ) echo "Please answer yes or no.";;
+    esac
+  done
+}
+
+fetch_plugins(){
+  while true; do
+    read -p "Would you like to install a(nother) plugin?" yn
+    case $yn in
         [Yy]* )
-          CURRENT=$1
-          WPTHEMES=($THEMES)
-          for THEME in "${WPTHEMES[@]}"; do
-            if [ $(basename $THEME) != $CURRENT ]; then
-              rm -rf $THEME
-            fi
-          done
+          read -p "Write your theme repository URL: " REPOSITORY
+          REPOSITORY=$1
+          PLUGINNAME=$(basename "$REPOSITORY")
+          PLUGINNAME="${THEMENAME%.*}" #removing extension
+          fetch_git $REPOSITORY $PLUGINNAME '/wp-content/plugins/'
+          read -p "Would you like to build '${PLUGINNAME}'?" yn
+          case $yn in
+            [Yy]* )
+              cd "${TARGETDIR}"
+              composer install --no-interaction $OPT_COMPOSER
+              npm install
+              npm run build --if-present
+              cd "${ROOT}"
+              break;;
+            [Nn]* ) exit;;
+            * ) echo "Please answer yes or no.";;
+          esac
           break;;
         [Nn]* ) exit;;
         * ) echo "Please answer yes or no.";;
@@ -84,20 +155,11 @@ npm install
 
 # SELECT THEME TO BUILD
 # =======================
-shopt -s nullglob
-WPTHEMES=('git repository URL');
-WPTHEMES+=($(ls -d $THEMES | xargs -n1 basename))
+fetch_theme
 
-echo "\nPlease select the theme to install:\n"
-select opt in "${WPTHEMES[@]}"
-do
-  if [[ $opt = 'git repository URL' ]]; then
-    read -p "Write your theme repository URL: " REPOSITORY
-    fetch_git $REPOSITORY
-  else
-    build_theme $opt
-  fi
-  break
-done
+# SELECT PLUGINS TO BUILD
+# =======================
+fetch_plugins
+
 echo  "${GREEN}Build finished.${NC}"
 exit
